+++ src/components/buttons/MobileMoreButton.tsx
@@ -5,7 +5,7 @@ import MoreHorizIcon from '@mui/icons-material/MoreHoriz';
 import Menu from '@mui/material/Menu';
 import MenuItem from '@mui/material/MenuItem';
 
-import { searchSongOnWeb, searchSongOnBili } from '../Fav/Favmenu';
+import { searchSongOnWeb, searchSongOnBili } from '../Fav/SongMenu';
 import { AddFavDialog } from '../dialogs/AddFavDialog';
 
 export default function mobileMoreButton({ song }: { song: NoxMedia.Song }) {
